+++ src/dropdown.js
@@ -81,9 +81,13 @@ class Dropdown extends EventEmitter {
    * @param {SearchResult[]} searchResults
    * @param {Dropdown~Offset} cursorOffset
    * @returns {this}
+   * @fires Dropdown#render
+   * @fires Dropdown#rendered
    */
   render(searchResults, cursorOffset) {
-    var rawResults = [], dropdownItems = [];
+    /** @event Dropdown#render */
+    this.emit('render');
+    let rawResults = [], dropdownItems = [];
     searchResults.forEach(searchResult => {
       rawResults.push(searchResult.data);
       if (dropdownItems.length < this.maxCount) {
@@ -93,8 +97,12 @@ class Dropdown extends EventEmitter {
     this.clear()
         .renderEdge(rawResults, 'header')
         .append(dropdownItems)
-        .renderEdge(rawResults, 'footer');
-    return this.items.length > 0 ? this.setOffset(cursorOffset).show() : this.hide();
+        .renderEdge(rawResults, 'footer')
+        .setOffset(cursorOffset)
+        .show();
+    /** @event Dropdown#rendered */
+    this.emit('rendered');
+    return this;
   }
 
   /**
@@ -201,7 +209,6 @@ class Dropdown extends EventEmitter {
    * @returns {this}
    * @fires Dropdown#show
    * @fires Dropdown#shown
-   * @fires Dropdown#rendered
    */
   show() {
     if (!this.shown) {
@@ -212,8 +219,6 @@ class Dropdown extends EventEmitter {
       /** @event Dropdown#shown */
       this.emit('shown');
     }
-    /** @event Dropdown#rendered */
-    this.emit('rendered');
     return this;
   }